Revert "Change image load check strategy"
Experimental revert of commit 
d6c142fd2bb8841dc9da5e2f15e25f23d1872e6f to see if it caused
the weirdness in the versus charts:
http://multimedia-metrics.wmflabs.org/dashboards/mmv#media_viewer_vs_file_page-graphs-tab

That commit affecting file page load times is extremely unlikely. 
A change in the test environment which exactly coincided with it is 
equally unlikely. This is an easy way  to tell which is the case.

Change-Id: I3581abd36d25494ca294db892480e25f6f4c5a73