Remove error report link
Took the easiest way and just removed it from the message as I don't
have time to work on this, no developer is looking at the error reports
and they are wasting Andre's time. Could be restored in the future with
more intelligent logic (to differentiate between network errors and
software bugs).

Bug: T111112
Change-Id: I3be22e5a821c32523aa15ac8cfb25baaaca4e477