Register configuration settings in the standard way
There's no need to do any fancy isset() on every setting - extension.json
will automatically take care of that.

For $wgMediaViewerEnableByDefaultForAnonymous which depended on the value
of $wgMediaViewerEnableByDefault, treat true/false as explicit values and
null as a fallback to the other configuration setting.

Keep the documentation associated with each setting in MultimediaViewer.php
for IDEs.

Change-Id: Ib34bbf75d027b8b96b2ec1a5975fa19061fadd55