Clear thumbs array when wikipage.content is run for the page content
Before this, all thumbs would be added to the array every time wikipage.content
is executed. This would result in an increasing amount of duplicated thumbs in
the array. Now, the array is cleared when the page content is updated and
therefore before the new thumbs are being processed, meaning that they will no
longer be duplicated.
Fixed version of patchset 1112294 which was reverted due to a regression.

Bug: T382520
Change-Id: I0d4cb0f73e8647b7b30d6f10e984129463fc83e0