Turn the metadata columns to single-column on <500px screens
This affected narrow windows showing desktop view at standard
resolution, and makes us more responsive to window and screen
sizes.

This will be important later if we replace the second partial
implementation of mobile multimedia viewer with the mainline
one in future, and benefits people using narrow desktop windows
immediately.

Note that I've left the 'More details' link where it is; on
narrow phones there's not much room for the title but it should
wrap correctly.

Bug: T378431
Change-Id: Ib9308855192b1cf03faa5ecb5ea0a2605882868e