Fix link being escaped due to security patch regression
Use .parseDom() (provided by jqueryMsg) to parse the system message
instead of escaping it. This way, we can insert the link into the
message as raw HTML and it is no longer escaped.
This also prevents unnecessary serialization/deserialization by
directly passing the jQuery object of the link to jqueryMsg and the
result of this to .append() instead of converting it to an HTML
string before passing it to jQuery.

Also insert the author parameter as HTML.

Bug: T397704
Change-Id: Ie5bc62ba9b4fe98b9cac8067ca621b7d55a1e5fe
(cherry picked from commit e9d33bc55f3b8a635be185a161073e8baca26b9f)